add progress card + dialog
